+++ CHANGELOG.md
@@ -1,2 +1,9 @@
 # v0.0.1 23/06/2025
 Inital Commit
+
+# v0.0.2 24/06/2025
+Added .gitignore, README.md and requirements.txt
+Set up folder structure
+Added test-program.py for use with pytest
+Added program.py with initial clear_folder function
+Added template PowerBI file in ./powerbi-renamer/test-pbi/